Pull logic from character presenter to service
