+++ app/models/character_search_text.rb
@@ -1,0 +1,21 @@
+module CharacterSearchText
+  def title
+    "MARVEL CHARACTERS"
+  end
+
+  def search_title
+    "SEARCH BY NAME"
+  end
+
+  def search_path
+    "/characters"
+  end
+
+  def search_results_title
+    "CHARACTER SEARCH RESULTS"
+  end
+
+  def search_placeholder
+    "CHARACTER NAME"
+  end
+end
